LPS-54824 match bookmarks
