+++ modules/apps/journal/journal-web/src/com/liferay/journal/web/customattributes/JournalFolderCustomAttributesDisplay.java
@@ -12,26 +12,37 @@
  * details.
  */
 
-package com.liferay.wiki.web;
+package com.liferay.journal.web.customattributes;
 
+import com.liferay.journal.web.constants.JournalPortletKeys;
 import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portlet.expando.model.BaseCustomAttributesDisplay;
-import com.liferay.wiki.model.WikiPage;
+import com.liferay.portlet.expando.model.CustomAttributesDisplay;
+import com.liferay.portlet.journal.model.JournalFolder;
+
+import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Jorge Ferrer
+ * @author Eudaldo Alonso
  */
-public class WikiPageCustomAttributesDisplay
+@Component(
+	immediate = true,
+	property = {
+		"javax.portlet.name=" + JournalPortletKeys.JOURNAL
+	},
+	service = CustomAttributesDisplay.class
+)
+public class JournalFolderCustomAttributesDisplay
 	extends BaseCustomAttributesDisplay {
 
 	@Override
 	public String getClassName() {
-		return WikiPage.class.getName();
+		return JournalFolder.class.getName();
 	}
 
 	@Override
 	public String getIconPath(ThemeDisplay themeDisplay) {
-		return themeDisplay.getPathThemeImages() + "/common/pages.png";
+		return themeDisplay.getPathThemeImages() + "/common/folder.png";
 	}
 
 }
